+++ common/src/main/java/pepjebs/mapatlases/client/screen/DecorationBookmarkButton.java
@@ -10,6 +10,8 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.world.level.saveddata.maps.MapDecoration;
 import net.minecraft.world.level.saveddata.maps.MapItemSavedData;
 import pepjebs.mapatlases.client.CompoundTooltip;
+import pepjebs.mapatlases.client.MapAtlasesClient;
+import pepjebs.mapatlases.config.MapAtlasesClientConfig;
 import pepjebs.mapatlases.integration.moonlight.CustomDecorationButton;
 import pepjebs.mapatlases.networking.C2SRemoveMarkerPacket;
 import pepjebs.mapatlases.networking.MapAtlasesNetworking;
@@ -138,10 +140,13 @@ public abstract class DecorationBookmarkButton extends BookmarkButton {
             return Tooltip.create(Component.translatable("tooltip.map_atlases.delete_marker"));
         }
         Component mapIconComponent = getDecorationName();
+        Tooltip t = Tooltip.create(mapIconComponent);
+        if (!MapAtlasesClientConfig.drawWorldMapCoords.get()) {
+            return t;
+        }
         Component coordsComponent = Component.literal("X: " + (int) getWorldX() + ", Z: " + (int) getWorldZ())
                 .withStyle(ChatFormatting.GRAY);
-        var t = Tooltip.create(mapIconComponent);
-        var t2 = Tooltip.create(coordsComponent);
+        Tooltip t2 = Tooltip.create(coordsComponent);
         return CompoundTooltip.create(t, t2);
     }