+++ changelog.md
@@ -1,2 +1,1 @@
-fixed a mixin type being incorrect
-increased max map count. i chalenge you to reach it without your game coming to a crawl
+- change map marker tooltip to not show coordinates when config is off